[2/2] CandyRoms: On-The-Go mode Squashed. Original pick from screwshop https://github.com/ScrewdAOSP/packages_apps_ScrewShop/commit/de8fdc8513bd7ed7e5b7b63dac640f09544a6c67
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

have fun!

Screw'd:
Bring up from MM sources
Squash in alpha settings by @LorDClockaN:
https://github.com/ScrewdAOSP/packages_apps_Settings/commit/ee5eced801813bf655c04679e805d0faef8a64f5
Modify alpha settings to use CustomSeekbarPreference

Change-Id: If1146cdb1529655dede774b371cc55f7b782f14e
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>